==> 0_ScrapeTweet.ipynb <==
import tweepy
import csv
import datetime
import numpy as np
import pandas as pd


#Twitter API credentials
consumer_key = "..."
consumer_secret = "..."
access_key = "..."
access_secret = "..."

auth = tweepy.OAuthHandler(consumer_key, consumer_secret)
auth.set_access_token(access_key, access_secret)
api = tweepy.API(auth)


Now we scrape the tweets that we need. At the moment we only get the last 30 tweets but another timeframe might be interesting: f.e. right before and after the European Election and then see if more similar before or afterwards

# Function to extract tweets 
def get_tweets(username, realname, party): 
          
        # Authorization to consumer key and consumer secret 
        auth = tweepy.OAuthHandler(consumer_key, consumer_secret) 
  
        # Access to user's access key and access secret 
        auth.set_access_token(access_key, access_secret) 
  
        # Calling api 
        api = tweepy.API(auth) 
  
        # 200 tweets to be extracted 
        tweets = api.user_timeline(screen_name=username, count = 200) 
        tweetsm = []
        
        for tweet in tweets:
          tweetsm.append(tweet.created_at)
  
        # Empty Array 
        MyEmptydf =  []
      
        # create array of tweet information: username,  
        # tweet id, date/time, text 
        tweets_for_csv = [tweet.text for tweet in tweets] # CSV file created  
        for j in tweets_for_csv: 
            
            # Appending tweets to the empty array tmp 
                MyEmptydf.append({ 'real': realname, 
                                   'tweet': j, 'user': username,
                                   'party': party,
                                 })
        # Printing the tweets 
        return(pd.DataFrame(MyEmptydf, tweetsm)) 
        
  
# Driver code 
if __name__ == '__main__': # what is this for???
  
    # Here goes the twitter handle for the user 
    # whose tweets are to be extracted. 
    get_tweets('akk', 'Annegreth Kramp-Karrenbauer', 'CDU')  

get_tweets("akk", "Annegreth Kramp-Karrenbauer", "CDU")  


# CDU
akk = get_tweets("akk","Annegreth Kramp-Karrenbauer", "CDU")   
merkel = get_tweets("AngelaMerkeICDU", "Angela Merkel", "CDU") 
# SPD
barley = get_tweets("katarinabarley", "Katharina Barley", "SPD") 
nahles = get_tweets("AndreaNahlesSPD", "Andrea Nahles", "SPD") 
# Grüne
oezdemir = get_tweets("cem_oezdemir", "Cem Özdemir", "Grüne") 
ekartd = get_tweets("GoeringEckardt", "Kathrin Goering-Eckardt", "Grüne") 
# AfD
weidel = get_tweets("Alice_Weidel", "Alice Weidel", "AfD")
meuthen = get_tweets("Joerg_Meuthen", "Jörg Meuthen", "AfD")
# Linke
wagenknecht = get_tweets("SWagenknecht", "Sarah Wagenknecht", "Linke")
# FDP
lindner = get_tweets("c_lindner", "Christian Lindner", "FDP")

frames = [akk, merkel, barley, nahles, oezdemir, ekartd, weidel, meuthen, wagenknecht, lindner]
df = pd.concat(frames, sort=True)

df

df.to_csv('/content/gdrive/My Drive/out.csv', sep=',', encoding='utf-8', index=True)
